+++ examples/redis/processor.php
@@ -4,14 +4,14 @@ declare(strict_types=1);
 use Tomaj\Hermes\Dispatcher;
 use Tomaj\Hermes\Driver\RedisSetDriver;
 use Tomaj\Hermes\Handler\EchoHandler;
-use Tomaj\Hermes\Restart\RedisRestart;
+use Tomaj\Hermes\Shutdown\RedisShutdown;
 
 require_once __DIR__.'/../../vendor/autoload.php';
 
 $redis = new Redis();
 $redis->connect('127.0.0.1', 6379);
 $driver = new RedisSetDriver($redis, 'hermes', 1);
-$driver->setRestart(new RedisRestart($redis));
+$driver->setShutdown(new RedisShutdown($redis));
 $driver->setupPriorityQueue('hermes_low', \Tomaj\Hermes\Dispatcher::PRIORITY_LOW);
 $driver->setupPriorityQueue('hermes_high', \Tomaj\Hermes\Dispatcher::PRIORITY_HIGH);