+++ src/Shutdown/ShutdownException.php
@@ -1,0 +1,10 @@
+<?php
+declare(strict_types=1);
+
+namespace Tomaj\Hermes\Shutdown;
+
+use Exception;
+
+class ShutdownException extends Exception
+{
+}